Safari-609.1.11.1/LayoutTests/fast/spatial-navigation/snav-offscreen-content-expected.txt

offscreen contentRight link

Down link
PASS gFocusedDocument.activeElement.getAttribute("id") is "1"
